+++ lib/cucumber/wire/protocol/requests.rb
@@ -1,5 +1,6 @@
+# coding: utf-8
 require 'cucumber/wire/request_handler'
-require 'cucumber/step_argument'
+require 'cucumber/wire/step_argument'
 
 module Cucumber
   module Wire
@@ -25,7 +26,7 @@ module Cucumber
           private
 
           def create_step_match(raw_step_match)
-            step_definition = StepDefinition.new(@connection, raw_step_match)
+            step_definition = StepDefinition.new(@connection, raw_step_match, @registry)
             step_args = raw_step_match['args'].map do |raw_arg|
               StepArgument.new(raw_arg['pos'], raw_arg['val'])
             end
@@ -69,7 +70,7 @@ module Cucumber
 
           def handle_diff!(tables)
             # TODO: figure out if / how we could get a location for a table from the wire (or make a null location)
-            location = Core::Ast::Location.new(__FILE__, __LINE__)
+            location = Core::Test::Location.new(__FILE__, __LINE__)
             table1 = table(tables[0], location)
             table2 = table(tables[1], location)
             table1.diff!(table2)
@@ -89,7 +90,7 @@ module Cucumber
           private
 
           def table(data, location)
-            Cucumber::MultilineArgument.from_core(Core::Ast::DataTable.new(data, location))
+            Cucumber::MultilineArgument.from_core(Core::Test::DataTable.new(data, location))
           end
         end